moved components to its directiories
