+++ ui/src/App.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import SignIn from './pages/SignIn';
 import HomePage from './pages/HomePage';
-import UserForm from './components/form/UserForm';
+import UserForm from './components/form/UserForm/UserForm';
 //const apiUrl = '/api';
 
 const App = () => {
